# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.CreateUserSwagger:
    properties:
      combat_certificate:
        example: true
        type: boolean
      disability_certificate:
        example:
        - ""
        items:
          type: string
        type: array
      email:
        example: example1@gmail.com
        type: string
      international_passport:
        example:
        - ""
        items:
          type: string
        type: array
      location:
        example: ""
        type: string
      name:
        example: ""
        type: string
      patronymic:
        example: ""
        type: string
      phone:
        example: "+380631122331"
        type: string
      physical_action_constraints:
        example:
        - ""
        items:
          type: string
        type: array
      surname:
        example: ""
        type: string
      weight_below_95:
        example: true
        type: boolean
    type: object
...
  model.Response:
    properties:
      code:
        type: integer
      messsage:
        type: string
    type: object
...
  model.UpdateUserSwagger:
    properties:
      combat_certificate:
        example: true
        type: boolean
      disability_certificate:
        example:
        - ""
        items:
          type: string
        type: array
      email:
        example: example1@gmail.com
        type: string
      id:
        example: ""
        type: string
      international_passport:
        example:
        - ""
        items:
          type: string
        type: array
      location:
        example: ""
        type: string
      name:
        example: ""
        type: string
      patronymic:
        example: ""
        type: string
      phone:
        example: "+380631122331"
        type: string
      physical_action_constraints:
        example:
        - ""
        items:
          type: string
        type: array
      surname:
        example: ""
        type: string
      weight_below_95:
        example: true
        type: boolean
    type: object
host: walking-school.site
info:
  contact:
    email: etverya12@gmail.com
    name: Yehor Tverytinov
  title: Walking-School backend API
  version: "1.0"
paths:
...
  /user:
    get:
      consumes:
      - application/json
      parameters:
      - description: limit
        in: query
        name: limit
        type: string
      - description: offset
        in: query
        name: offset
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get all users.
      tags:
      - user
    post:
      consumes:
      - application/json
      parameters:
      - description: User
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/model.CreateUserSwagger'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Create user.
      tags:
      - user
    put:
      consumes:
      - application/json
      parameters:
      - description: User
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Update user by id.
      tags:
      - user
  /user/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Delete user by id.
      tags:
      - user
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get user by id.
      tags:
      - user
swagger: "2.0"